refactor(epicenter): improve static workspace architecture
- Move route-specific components to _components/ folder
- Load initial table/KV data in +layout.ts instead of $effect
- Fix Y.Doc discovery by forcing type instantiation with getArray()
- Simplify viewer components to use initialData props
- Remove debug console.log statements
