+++ apps/epicenter/src/routes/(workspace)/workspaces/static/[id]/+layout.ts
@@ -4,7 +4,12 @@ import {
 	createYSweetConnection,
 	getDefaultSyncUrl,
 } from '$lib/docs/y-sweet-connection';
-import { discoverKvKeys, discoverTables } from '$lib/docs/discover';
+import {
+	discoverKvKeys,
+	discoverTables,
+	readAllKv,
+	readTableRows,
+} from '$lib/docs/discover';
 
 /**
  * Load a static workspace by ID.
@@ -17,7 +22,6 @@ import { discoverKvKeys, discoverTables } from '$lib/docs/discover';
  */
 export const load: LayoutLoad = async ({ params }) => {
 	const workspaceId = params.id;
-	console.log(`[StaticLayout] Loading static workspace: ${workspaceId}`);
 
 	// Get registry entry (may not exist for ad-hoc viewing)
 	const entry = await getStaticWorkspace(workspaceId);
@@ -25,15 +29,11 @@ export const load: LayoutLoad = async ({ params }) => {
 	// Determine sync URL
 	const syncUrl = entry?.syncUrl ?? getDefaultSyncUrl();
 
-	console.log(`[StaticLayout] Connecting to Y-Sweet at: ${syncUrl}`);
-
 	// Create Y-Sweet connection
 	const connection = createYSweetConnection({
 		workspaceId,
 		serverUrl: syncUrl,
 	});
-
-	console.log(`[StaticLayout] Provider created, waiting for sync...`);
 
 	// Wait for initial sync with timeout
 	let syncStatus = 'unknown';
@@ -45,43 +45,23 @@ export const load: LayoutLoad = async ({ params }) => {
 			),
 		]);
 		syncStatus = 'synced';
-		console.log(`[StaticLayout] Sync complete!`);
-	} catch (e) {
+	} catch {
 		syncStatus = 'failed';
-		console.error(`[StaticLayout] Failed to sync: ${e}`);
 		// Don't throw - allow viewing even if sync fails
-	}
-
-	// Debug: Log raw Y.Doc state
-	console.log(`[StaticLayout] Y.Doc guid: ${connection.ydoc.guid}`);
-	console.log(
-		`[StaticLayout] Y.Doc share keys:`,
-		[...connection.ydoc.share.keys()],
-	);
-
-	// Debug: Log type information for each share key
-	// Force instantiation using getArray() to see actual data
-	for (const key of connection.ydoc.share.keys()) {
-		if (key.startsWith('table:')) {
-			const array = connection.ydoc.getArray(key);
-			console.log(`[StaticLayout] Share key '${key}' (via getArray):`, {
-				constructor: array?.constructor?.name,
-				length: array.length,
-				firstTwo: array.toArray().slice(0, 2),
-			});
-		}
 	}
 
 	// Discover structure
 	const tables = discoverTables(connection.ydoc);
 	const kvKeys = discoverKvKeys(connection.ydoc);
 
-	console.log(
-		`[StaticLayout] Discovered ${tables.length} tables, ${kvKeys.length} KV keys`,
-	);
-	console.log(`[StaticLayout] Tables:`, tables);
-	console.log(`[StaticLayout] KV Keys:`, kvKeys);
-	console.log(`[StaticLayout] Sync status: ${syncStatus}`);
+	// Read initial data for all tables
+	const tableData: Record<string, Record<string, unknown>[]> = {};
+	for (const tableName of tables) {
+		tableData[tableName] = readTableRows(connection.ydoc, tableName);
+	}
+
+	// Read all KV values
+	const kvData = readAllKv(connection.ydoc);
 
 	return {
 		workspaceId,
@@ -89,6 +69,9 @@ export const load: LayoutLoad = async ({ params }) => {
 		connection,
 		tables,
 		kvKeys,
+		tableData,
+		kvData,
+		syncStatus,
 		displayName: entry?.name ?? workspaceId,
 	};
 };
