+++ apps/epicenter/src/routes/(workspace)/workspaces/static/[id]/_components/StaticWorkspaceSidebar.svelte
